Move to sibling-unique IDs instead of global keys
This is actually a rather excessively large commit covering a lot of
territory, but the core thing it does is move GReact away from requiring
every node to have a globally unique key. This provided a significant
burden when building large and complicated UI graphs, and required a lot
of string composition overhead every frame. The only real advantage is
that it made it easy to handle nodes maintaining their identity while
moving to different parts of the scene graph. As this is a pretty rare
occurrance, this trade-off was deemed not actually worth it.

Instead, GReact now uses int IDs, which only need to be unique among
siblings of the same node type, which are used to maintain node identity
between frames. This means identity can only actually be preserved as
long as the node stays in the same place in the scene graph. By default,
GReact will generate these IDs sequentially starting at 0 for the all
sibling nodes of a given type. However, all node component props structs
must now have a field called "id" of type "int?", and if that's present
it will override the normal method of generating IDs.

In order to make this work, the core render algorithm has been totally
rewritten. Where before it shoved everything from the old graph and the
new graph into a set of dictionaries, then iterated over them to build
queues of elements to be created, destroyed, moved or updated, now it
recursively navigates the graph and creates, destroys or updates nodes
as it goes. This is actually much simpler.

Additional changes to the GReact core:
- There is also now an interface called INodeProps, and Element now
requires that the prop type given to it implement INodeProps. This is to
enforce the presence of the id property.
- Element also now has a new NodeT generic type argument, which is the
type of the node it creates, and a matching nodeType property that
returns the Type of that type argument.
- Element.Render no longer supports the node type changing, and instead
throws an exception when that handles. Given how IDs are now handled, it
should be impossible for the old element passed in to be a different
type.
- PopulatedElement now tracks it's children in a dictionary indexed by
id, instead of just an array, where the id is a tuple of (Type, int).
- PopulatedElement also keeps a dictionary of the maximum id assigned to
any child by type, in a dictionary, to avoid auto-assigned ids from
having collisions.
- PopulatedElement now has a GetNextId method that takes a type, and
returns the next available id for that type. It assumes that id is being
assigned to a node, and increments the maximum assigned id.
- Signal no longer implements IEquatable<Signal>, but instead overrides
the Equals method from object, because it turns out IEquatable is not
used when comparing the fields of a struct, but object.Equals is. This
also required adding a GetHashCode implementation. This was due to a bug
uncovered by the updated UI example.
- Added a Connect method to Signal, for registering the signal on a
Godot node. It takes the old signal, so it can do the comparison itself
to only update it if necessary, and to unregister the old one when it
does.

Additional changes to GReact node components:
- Added LabelComponent, because it was needed by the updated UI example.
- IControlProps now has minSize, sizeFlagsHoriz and sizeFlagsVert to
provide the controls necessary to format a control that's inside a
container. Add handling of these props to ControlComponent.CopyToNode.
This was added here because the updated UI example made heavier use of
containers.
- ControlComponent.CopyToNode now checks if vert or horiz have changed
before assigning them, which prevents it from overwriting the values
assigned by a container. This requires taking the old props as an
argument, which means all controls with a CopyToNode now must do that.
- BaseButton now uses the new Signal.Connect call.

Changes to the Examples:
- The UI example was replaced with a much more extensive one. It now
builds a UI with a sequence of lists, where lists can be added or
removed, and elements can be added to removed from the lists. This was
to more thuroughly test out the new id behavior.