Add ability for Renderer to track node churn
The Renderer now has a public bool on it called trackNodeChurn, along
with two read-only ints called nodesCreated and nodesDestroyed. If
trackNodeChurn is set to true, then each time it renders, it'll set
nodesCreated and nodesDestroyed to the number of Godot nodes it creates
and destroys during that render. If it's false, they'll always just be
set to zero. The reason for the toggle is that tracking the nodes
destroyed requires doing a recrusive crawl of the children of the
destroyed node to count them all, which could potentially be a little
bit slow, so best to only do this during normal operation.