Add LineEditComponent
Also updated the UI example to use the LineEdit component.

Unlike most node components, the LineEdit component actually reads a
value on the node in it's ModifyNode call. This is because changing the
"Text" member on LineEdit will move the carrot back to the beginning,
disrupting the user's typing, but the valye of "Text" can be changed by
direct user input, not just by GReact. The combination of these factors
means that just checking against the oldProps to decide whether to
change the value isn't sufficient, as it will often appear to require
an update when it doesn't, and updating it anyway will be disruptive.