Refactor and improve the UI example
- Replace all usage of "column" with "list".
- Simplify the state, removing the Column class and instead just using a
List<string> directly.
- Add an action system to the state. It now has a set of static
methods that return functions that take a State and modify it, and has
an Apply method that takes such a function and calls it, so that the
Apply method can be passed in to components without them needing to have
all the parts of the state they may need to modify in response to a
signal. This gives a better sense of the flow of a proper state system
for use with GReact.
- Tweak all the static methods used for signals to follow are more
consistent pattern. They now all take the component's props struct,
called "props", or take a tuple of the props struct and one other value,
called "args", which is destructured on the first line.