Signals are now given their node as an arg
Every signal is now given the node it's attached to as the first
argument to it's callback. This was inspired by the desire for a new
control to grab focus when it's created. Godot manages focus quite well,
and the only way for the GReact paradigm to directly manipulate focus
would require basically entirely re-implementing it, which wouldn't be
desirable. This allows you to create an onReady signal that just calls
GrabFocus on the node. Making it a generalized thing for all signal
types means it's a consistent pattern that should be able to handle many
future use cases.

This also required adding the onReady signal to IControlProps. This may
eventually need to be moved up to INodeProps.

Then updated the UI example to make use of this to focus and select all
on newly created LineEdit nodes.