Make UIDim also manage container positioning
Godot has two ways of positioning controls in it's UI: you can do so
manually with anchors and margins, or, if the control is inside a
container, you can use the minSize and size flags. Previously, UIDim
was only used for the manual positioning, and control also had size
flags and minSize separately. Now, they're combined, and UIDim manages
all of these, which through the use of it's many convenient constructors
should make it much clearer what you need to do to format a control
correctly.