Add source generator for cleaner API
This commit adds a C# 9 source generator, which creates a bunch of
automatic boilerplate at compile time to allow for a cleaner API.

Specifically, it creates a static class called Component, and adds a
static function to it for each component in the codebase, both GReact
components and custom components. These functions will have an optional
named argument for each prop, so you no longer have to explicitly make
the props struct.

To enable this, there are now much stricter requirements for exactly how
a component needs to be structured. A lot of things that were previously
just conventions are now requirements. The component must be a static
function called "New", on a static class. That static class must have
the [Component] attribute on it, and the props struct must be inside the
the static class, and it must be called "Props".

Itemized changes:
- The generator is it's own project, in the GReactGenerator directory.
- The GReact.csproj project can no longer use default compile items,
because it has to exclude GReactGenerator. So instead, it just
explicitly includes the GReact and Examples directories. It also
includes the DLL created by the generator project as an analyzer.
- The GReact.sln adds the GReactGenerator project.
- Added the Chilren function on Element, which is like Child but
variadic, to add an arbitrary number of children at once.
- Changed many "new" expressions to use the new C# 9 syntax which no
longer redundantly specifies the type.
- Added the "Component" attribute to GReact.
- Updated all the GReact node components to follow the new pattern.
- Updated the UI example to use the new API.
- Updated the README and overview to reflect the new API.
- Added .vscode/tasks.json and settings.json. The tasks adds a build
task for the generator DLL, and the settings indicates usage of tabs,
and format on paste and save.