Fix a crash when the root node changes type
In most circumstances, when a node is replaced with another node of a
different type, it would have a different key, and thus be considered a
new node. So the old node would be freed and a new one created. However,
it followed a slightly different code path for the root node which
inadvertently skipped this check. That's now properly handled, so it
will no longer crash if the root node changes type.