Redesign signal API to use simple closures
Signals can now be passed in as simple closures, instead of using the
Signal class. To make this work requires expanding the source generator.

For each type of node component that requires signals, it creates a
Signals class. For each signal type that node component has, the Signals
class will have a closure member variable to store to current registered
callback, and a member variable that calls that closure.

For each such node component, it will also create a registration
function, called RegisterFooSignals, where "Foo" is the name of the node
component. This function takes the node, and the node's props, and for
each prop representing a signal, it updates closure on the Signals
object. If the closure becomes null, or becomes not-null, it will
connect or disconnect the Godot signal, as necessary. This function must
be called in the create and modify functions of any node component with
signals. Which should be all node components, since every node at least
has the "ready" signal.

The node component specifies signals with the new SignalAttribute, which
must be placed on each closure prop that's intended to correspond to a
signal on the node. It takes one argument, which is the name of the
Godot signal it corresponds to.